another component added
