+++ src/Components/AlertBox/AlertBox.scss
@@ -1,29 +1,13 @@
-.AlertItem {
-  display: flex;
-  width: 100%;
-  align-items: center;
-  justify-content: flex-end;
-  top: 0;
-  position: sticky;
-  &--hidden {
-    display: none;
-  }
-  &__text {
-    width: 95%;
-    text-align: center;
-  }
-  &__space {
-    width: 5%;
-    text-align: right;
-    margin-right: 10px;
-  }
-}
-
-.bottomContainer {
+.alertContainer {
   display: flex;
   width: 100%;
   flex-direction: column;
-  position: absolute;
-  top: auto;
-  bottom: 10px;
+  top: 0;
+  position: sticky;
+  &--bottom {
+    width: 100%;
+    position: absolute;
+    top: auto;
+    bottom: 2px;
+  }
 }
