+++ src/Components/index.js
@@ -1,4 +1,4 @@
-import AlertBox from "../Components/AlertBox/index";
+import { AlertBox } from "../Components/AlertBox/index";
 import LoginPanel from "../Components/LoginPanel/index";
 import ShopItem from "../Components/ShopItem/index";
 import { Menu } from "../Components/SidebarMenu/index";
